Simplify package structure.
