+++ JavaModules/idispatch-client/src/main/java/net/pkhapps/idispatch/client/v3/AssignmentStateId.java
@@ -1,4 +1,4 @@
-package net.pkhapps.idispatch.client.v3.type;
+package net.pkhapps.idispatch.client.v3;
 
 import net.pkhapps.idispatch.client.v3.base.DomainObjectId;
 
@@ -6,11 +6,11 @@ import javax.annotation.Nonnull;
 import javax.annotation.concurrent.Immutable;
 
 /**
- * ID type for {@link AssignmentType}.
+ * ID type for {@link AssignmentState}.
  */
 @Immutable
-public class AssignmentTypeId extends DomainObjectId {
-    public AssignmentTypeId(@Nonnull Long id) {
+public class AssignmentStateId extends DomainObjectId {
+    public AssignmentStateId(@Nonnull Long id) {
         super(id);
     }
 }
