+++ JavaModules/idispatch-client/src/main/java/net/pkhapps/idispatch/client/v3/AssignmentTypeId.java
@@ -1,0 +1,16 @@
+package net.pkhapps.idispatch.client.v3;
+
+import net.pkhapps.idispatch.client.v3.base.DomainObjectId;
+
+import javax.annotation.Nonnull;
+import javax.annotation.concurrent.Immutable;
+
+/**
+ * ID type for {@link AssignmentType}.
+ */
+@Immutable
+public class AssignmentTypeId extends DomainObjectId {
+    public AssignmentTypeId(@Nonnull Long id) {
+        super(id);
+    }
+}
