+++ JavaModules/idispatch-client/src/main/java/net/pkhapps/idispatch/client/v3/Municipality.java
@@ -1,4 +1,4 @@
-package net.pkhapps.idispatch.client.v3.type;
+package net.pkhapps.idispatch.client.v3;
 
 import net.pkhapps.idispatch.client.v3.base.DeactivatableDomainObject;
 import net.pkhapps.idispatch.client.v3.base.IdentifiableDomainObject;
@@ -9,17 +9,17 @@ import javax.annotation.concurrent.Immutable;
 import java.util.Objects;
 
 /**
- * Identifiable value object representing a resource type.
+ * Identifiable value object representing a municipality.
  */
 @Immutable
 @SuppressWarnings("WeakerAccess")
-public class ResourceType implements IdentifiableDomainObject<ResourceTypeId>, DeactivatableDomainObject {
+public class Municipality implements IdentifiableDomainObject<MunicipalityId>, DeactivatableDomainObject {
 
-    private ResourceTypeId id;
+    private MunicipalityId id;
     private MultilingualString name;
     private boolean active;
 
-    public ResourceType(@Nonnull ResourceTypeId id, @Nonnull MultilingualString name, boolean active) {
+    public Municipality(@Nonnull MunicipalityId id, @Nonnull MultilingualString name, boolean active) {
         this.id = Objects.requireNonNull(id, "id must not be null");
         this.name = Objects.requireNonNull(name, "name must not be null");
         this.active = active;
@@ -27,7 +27,7 @@ public class ResourceType implements IdentifiableDomainObject<ResourceTypeId>, D
 
     @Nonnull
     @Override
-    public ResourceTypeId id() {
+    public MunicipalityId id() {
         return id;
     }
 
@@ -45,7 +45,7 @@ public class ResourceType implements IdentifiableDomainObject<ResourceTypeId>, D
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        ResourceType that = (ResourceType) o;
+        Municipality that = (Municipality) o;
         return active == that.active &&
                 Objects.equals(id, that.id) &&
                 Objects.equals(name, that.name);
@@ -58,6 +58,6 @@ public class ResourceType implements IdentifiableDomainObject<ResourceTypeId>, D
 
     @Override
     public String toString() {
-        return String.format("%s[id=%s, active=%s]", getClass().getSimpleName(), id, active);
+        return String.format("%s[id=%s, name=%s, active=%s]", getClass().getSimpleName(), id, name, active);
     }
 }
