+++ JavaModules/idispatch-client/src/main/java/net/pkhapps/idispatch/client/v3/MunicipalityId.java
@@ -1,4 +1,4 @@
-package net.pkhapps.idispatch.client.v3.type;
+package net.pkhapps.idispatch.client.v3;
 
 import net.pkhapps.idispatch.client.v3.base.DomainObjectId;
 
@@ -6,11 +6,11 @@ import javax.annotation.Nonnull;
 import javax.annotation.concurrent.Immutable;
 
 /**
- * ID type for {@link ResourceType}.
+ * ID type for {@link Municipality}.
  */
 @Immutable
-public class ResourceTypeId extends DomainObjectId {
-    public ResourceTypeId(@Nonnull Long id) {
+public class MunicipalityId extends DomainObjectId {
+    public MunicipalityId(@Nonnull Long id) {
         super(id);
     }
 }
