+++ JavaModules/idispatch-client/src/main/java/net/pkhapps/idispatch/client/v3/ResourceId.java
@@ -1,4 +1,4 @@
-package net.pkhapps.idispatch.client.v3.type;
+package net.pkhapps.idispatch.client.v3;
 
 import net.pkhapps.idispatch.client.v3.base.DomainObjectId;
 
@@ -6,11 +6,11 @@ import javax.annotation.Nonnull;
 import javax.annotation.concurrent.Immutable;
 
 /**
- * ID type for {@link Municipality}.
+ * ID type for {@link Resource}.
  */
 @Immutable
-public class MunicipalityId extends DomainObjectId {
-    public MunicipalityId(@Nonnull Long id) {
+public class ResourceId extends DomainObjectId {
+    public ResourceId(@Nonnull Long id) {
         super(id);
     }
 }
