+++ JavaModules/idispatch-client/src/main/java/net/pkhapps/idispatch/client/v3/ResourceStateLookupService.java
@@ -1,4 +1,4 @@
-package net.pkhapps.idispatch.client.v3.type;
+package net.pkhapps.idispatch.client.v3;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
