+++ JavaModules/idispatch-client/src/main/java/net/pkhapps/idispatch/client/v3/ResourceTypeId.java
@@ -1,4 +1,4 @@
-package net.pkhapps.idispatch.client.v3.type;
+package net.pkhapps.idispatch.client.v3;
 
 import net.pkhapps.idispatch.client.v3.base.DomainObjectId;
 
@@ -6,11 +6,11 @@ import javax.annotation.Nonnull;
 import javax.annotation.concurrent.Immutable;
 
 /**
- * ID type for {@link ResourceState}.
+ * ID type for {@link ResourceType}.
  */
 @Immutable
-public class ResourceStateId extends DomainObjectId {
-    public ResourceStateId(@Nonnull Long id) {
+public class ResourceTypeId extends DomainObjectId {
+    public ResourceTypeId(@Nonnull Long id) {
         super(id);
     }
 }
