+++ JavaModules/idispatch-client/src/main/java/net/pkhapps/idispatch/client/v3/StationId.java
@@ -1,4 +1,4 @@
-package net.pkhapps.idispatch.client.v3.type;
+package net.pkhapps.idispatch.client.v3;
 
 import net.pkhapps.idispatch.client.v3.base.DomainObjectId;
 
@@ -6,11 +6,11 @@ import javax.annotation.Nonnull;
 import javax.annotation.concurrent.Immutable;
 
 /**
- * ID type for {@link Resource}.
+ * ID type for {@link Station}.
  */
 @Immutable
-public class ResourceId extends DomainObjectId {
-    public ResourceId(@Nonnull Long id) {
+public class StationId extends DomainObjectId {
+    public StationId(@Nonnull Long id) {
         super(id);
     }
 }
