+++ JavaModules/idispatch-client/src/test/java/net/pkhapps/idispatch/client/v3/MunicipalityTest.java
@@ -1,5 +1,7 @@
-package net.pkhapps.idispatch.client.v3.type;
+package net.pkhapps.idispatch.client.v3;
 
+import net.pkhapps.idispatch.client.v3.Municipality;
+import net.pkhapps.idispatch.client.v3.MunicipalityId;
 import net.pkhapps.idispatch.client.v3.test.JsonObjectTest;
 import net.pkhapps.idispatch.client.v3.util.MultilingualString;
 import org.junit.Test;
@@ -9,26 +11,26 @@ import static net.pkhapps.idispatch.client.v3.util.Locales.SWEDISH;
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
- * Unit test for {@link ResourceType}.
+ * Unit test for {@link Municipality}.
  */
 @SuppressWarnings("WeakerAccess")
-public class ResourceTypeTest extends JsonObjectTest {
+public class MunicipalityTest extends JsonObjectTest {
 
-    public static ResourceType createTestResourceType() {
-        return new ResourceType(new ResourceTypeId(1L),
+    public static Municipality createTestMunicipality() {
+        return new Municipality(new MunicipalityId(1L),
                 new MultilingualString.Builder()
-                        .withValue(SWEDISH, "Räddningsenhet")
-                        .withValue(FINNISH, "Pelastusyksikkö")
+                        .withValue(SWEDISH, "Testkommun")
+                        .withValue(FINNISH, "Testikunta")
                         .build(),
                 true);
     }
 
     @Test
     public void serializeAndDeserialize() {
-        var original = createTestResourceType();
+        var original = createTestMunicipality();
         var json = getGson().toJson(original);
         System.out.println(json);
-        var deserialized = getGson().fromJson(json, ResourceType.class);
+        var deserialized = getGson().fromJson(json, Municipality.class);
         assertThat(deserialized).isNotSameAs(original);
         assertThat(deserialized).isEqualTo(original);
     }
