+++ JavaModules/idispatch-client/src/test/java/net/pkhapps/idispatch/client/v3/ResourceTest.java
@@ -1,4 +1,4 @@
-package net.pkhapps.idispatch.client.v3.type;
+package net.pkhapps.idispatch.client.v3;
 
 import net.pkhapps.idispatch.client.v3.test.JsonObjectTest;
 import org.junit.Test;
