+++ JavaModules/idispatch-client/src/test/java/net/pkhapps/idispatch/client/v3/ResourceTypeTest.java
@@ -1,5 +1,7 @@
-package net.pkhapps.idispatch.client.v3.type;
+package net.pkhapps.idispatch.client.v3;
 
+import net.pkhapps.idispatch.client.v3.ResourceType;
+import net.pkhapps.idispatch.client.v3.ResourceTypeId;
 import net.pkhapps.idispatch.client.v3.test.JsonObjectTest;
 import net.pkhapps.idispatch.client.v3.util.MultilingualString;
 import org.junit.Test;
@@ -9,26 +11,26 @@ import static net.pkhapps.idispatch.client.v3.util.Locales.SWEDISH;
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
- * Unit test for {@link Municipality}.
+ * Unit test for {@link ResourceType}.
  */
 @SuppressWarnings("WeakerAccess")
-public class MunicipalityTest extends JsonObjectTest {
+public class ResourceTypeTest extends JsonObjectTest {
 
-    public static Municipality createTestMunicipality() {
-        return new Municipality(new MunicipalityId(1L),
+    public static ResourceType createTestResourceType() {
+        return new ResourceType(new ResourceTypeId(1L),
                 new MultilingualString.Builder()
-                        .withValue(SWEDISH, "Testkommun")
-                        .withValue(FINNISH, "Testikunta")
+                        .withValue(SWEDISH, "Räddningsenhet")
+                        .withValue(FINNISH, "Pelastusyksikkö")
                         .build(),
                 true);
     }
 
     @Test
     public void serializeAndDeserialize() {
-        var original = createTestMunicipality();
+        var original = createTestResourceType();
         var json = getGson().toJson(original);
         System.out.println(json);
-        var deserialized = getGson().fromJson(json, Municipality.class);
+        var deserialized = getGson().fromJson(json, ResourceType.class);
         assertThat(deserialized).isNotSameAs(original);
         assertThat(deserialized).isEqualTo(original);
     }
